Localization finifhed, added new service
Implementation of the localization service is done. The strings are now
fetched from the object in the languages.js file.

The continiued refactoring of the "muddy" controller in the logic.js
file led me to implement a new service witch is responsible for creating
the constants used in calculations.
Its better design to maintain all of them in one place, and it helps
"cleaning" the controller.

Comments:
I´ve added several comments and "java"doc for some of the functions and
variables.